# Conflicts:
#	Source/Mame/Scene/SceneGame.cpp
#	imgui.ini

+++ Source/Taki174/Common.h
@@ -43,10 +43,8 @@ inline void rotate(
     y += cy;
 };
 
-// float版rand関数(※事前にsrand処理が必要)
-inline const float RandFloat(
-    const float min,
-    const float max)
+// ランダムfloat関数(※事前にsrand処理が必要)
+inline const float RandFloat(const float min, const float max)
 {
     // 0.0～1.0の間までのランダム値
     constexpr float randMaxf = 0x7fff;
@@ -56,9 +54,8 @@ inline const float RandFloat(
     return value * (max - min) + min;
 }
 
-inline const float RandInt(
-    const int min,
-    const int max)
+// ランダムint関数(※事前にsrand処理が必要)
+inline const int RandInt(const int min, const int max)
 {
     return ::rand() % ((max + 1) - min) + min;
 }
